Time: 177 ms (70.25%), Space: 45.2 MB (70.75%) - LeetHub

+++ 0992-subarrays-with-k-different-integers/0992-subarrays-with-k-different-integers.cpp
@@ -2,7 +2,7 @@ class Solution {
 public:
     int atmost(vector<int> &nums,int k){
         int n=nums.size();
-        int l=0,r=0;
+        long long int l=0,r=0;
         int size=0,cnt=0;
         unordered_map<int,int> mpp;
         while(r<n){
